requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
eval==0.0.5
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SSE==1.0.0
git-filter-repo==2.47.0
gunicorn==20.0.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pymongo==4.10.1
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.1
setuptools==75.6.0
six==1.17.0
Werkzeug==3.1.3
wheel==0.45.1
